drm/vc4: hdmi: Simplify the hotplug handling
BugLink: https://bugs.launchpad.net/bugs/1995968

Our detect callback has a bunch of operations to perform depending on
the current and last status of the connector, such a setting the CEC
physical address or enabling the scrambling again.

This is currently dealt with a bunch of if / else statetements that make
it fairly difficult to read and extend.

Let's move all that logic to a function of its own.

Signed-off-by: Maxime Ripard <maxime@cerno.tech>

(cherry picked from commit 53dde128b849b04057e29b3083fb933e75c81fd8 rpi-5.19.y)
Signed-off-by: Juerg Haefliger <juerg.haefliger@canonical.com>
Acked-by: Cory Todd <cory.todd@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>
Signed-off-by: Juerg Haefliger <juerg.haefliger@canonical.com>