media: i2c: ov9281: Correct min/def vts for 640x400
BugLink: https://bugs.launchpad.net/bugs/1982090

A VTS of 421 in the 640x400 mode only streams at 130.2fps
instead of the expected ~261fps. This appears to be an invalid
value for the sensor, as a VTS of 422 works fine to give 259.79fps.

Set the minimum (and default) to 422.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>

(cherry picked from commit 0a36922a541a5a71d6bdcdbd7e09379bc93a2ef4 rpi-5.19.y)
Signed-off-by: Juerg Haefliger <juerg.haefliger@canonical.com>